fix(review): centralize audit contract validators
