+++ src/autoskillit/core/types/_type_audit_admission_ledger.py
@@ -15,7 +15,6 @@ from enum import StrEnum
 from pathlib import Path
 from typing import Any, ClassVar, Literal, Protocol, runtime_checkable
 
-from ..closure_hashing import HASH_RE
 from ._type_audit_admission import (
     AuditAttemptId,
     AuditIdentityReservation,
@@ -27,6 +26,12 @@ from ._type_audit_admission import (
     ReservationDecision,
 )
 from ._type_audit_cycle import ArtifactRef, AuditCycleHead
+from ._type_audit_validation import (
+    _require_absolute_path,
+    _require_digest,
+    _require_nonempty,
+    _require_optional_digest,
+)
 
 __all__ = [
     "AuditAdmissionLedger",
@@ -46,30 +51,6 @@ __all__ = [
     "AuditReservationOutcome",
     "AuditReservationRequest",
 ]
-
-
-def _require_nonempty(name: str, value: object) -> str:
-    if not isinstance(value, str) or not value.strip():
-        raise ValueError(f"{name} must be a non-empty string")
-    return value
-
-
-def _require_digest(name: str, value: object) -> str:
-    if not isinstance(value, str) or HASH_RE.fullmatch(value) is None:
-        raise ValueError(f"{name} must be an algorithm-qualified sha256 digest")
-    return value
-
-
-def _require_optional_digest(name: str, value: object) -> str | None:
-    if value is None:
-        return None
-    return _require_digest(name, value)
-
-
-def _require_absolute_path(name: str, value: object) -> Path:
-    if not isinstance(value, Path) or not value.is_absolute() or ".." in value.parts:
-        raise ValueError(f"{name} must be an absolute non-traversing Path")
-    return value
 
 
 class AuditAdmissionStorageHealthStatus(StrEnum):
